requirements.txt
openai
pyaudio
requests
wave
playsound
python-dotenv
pynput